+++ packages/terra-functional-testing/src/terra-cli/wdio/index.js
@@ -5,6 +5,10 @@ const cli = {
   describe: 'Run wdio tests',
   builder: (yargs) => (
     yargs.options({
+      browsers: {
+        type: 'array',
+        describe: 'A list of browsers for the test run.',
+      },
       c: {
         type: 'string',
         alias: 'config',
@@ -13,6 +17,15 @@ const cli = {
       formFactors: {
         type: 'array',
         describe: 'A list of form factors for the test run.',
+      },
+      gridUrl: {
+        type: 'string',
+        describe: 'The remote selenium grid address.',
+      },
+      keepAliveSeleniumDockerService: {
+        type: 'boolean',
+        describe: 'Determines to keep the selenium docker service running upon test completion.',
+        default: false,
       },
       locales: {
         type: 'array',
@@ -43,11 +56,6 @@ const cli = {
       spec: {
         type: 'array',
         describe: 'A list of spec file paths.',
-      },
-      keepAliveSeleniumDockerService: {
-        type: 'boolean',
-        describe: 'Determines to keep the selenium docker service running upon test completion.',
-        default: false,
       },
     })
   ),