+++ packages/terra-toolkit-docs/CHANGELOG.md
@@ -5,6 +5,7 @@
 * Changed
   * Update webpack-config-terra documentation.
   * Add stylelint-config-terra documentation
+  * Update webpack-config-terra upgrade guide.
 
 ## 1.3.0 - (December 7, 2020)